Rename 'implementation' to 'tree_node'
